refactor: Calculator의 명칭을 CalculatorService로 정정한다
- 클래스 명을 CalculatorService로 정정한다
- service 폴더를 추가 및 CalculatorService를 이동시킨다
