+++ src/main/java/calculator/Application.java
@@ -1,8 +1,10 @@
 package calculator;
 
+import calculator.service.CalculatorService;
+
 public class Application {
 	public static void main(String[] args) {
-		Calculator calculator = new Calculator();
+		CalculatorService calculator = new CalculatorService();
 		calculator.run();
 	}
 }
